FingerprintOffline siger nu en lyd, når den har indsamlet alle measurements. ExperimentActivity kan nu pauses og startes igen.

+++ ShittyApp/app/src/main/java/com/example/henneberg/shittyapp/Util/AppConstants.java
@@ -187,7 +187,7 @@ public class AppConstants extends AppCompatActivity {
         try {
             obj.put("value", (int) RSSI);
             obj.put("phoneID", Integer.parseInt(getPhoneName()));
-            obj.put("time", timestamp);
+            obj.put("time", (int) timestamp);
             obj.put("run", getExperimentRun());
             //Phone ID, number
             //Run (experiment run), number
